backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
SQLAlchemy
tensorflow
joblib
pandas
numpy
bcrypt
cryptography
python-dotenv
scikit-learn
Werkzeug
PyMySQL
